Use EVTCHNOP_unmask to unmask in some cases
If the EVTCHN Unmask method is called in context of the event channel
upcall, or on a CPU to which the event is not bound then use the unmask
hypercall. In the former case not doing this may lead to priorty
inversion once setting event channel priorities is enabled. In the latter
case not doing it again stops a lower priority event handler trigger from
blocking a higher priority event injection.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>