Re-instate bus enumation filtering in XENFILT
The filtering code was removed in commit
ff034f7ebd4010f7b502c56cb1cca6ba40a7f1aa when it was realized that
simply modifying the reported instance ID of the active device to be
constant, even if it moves location on the PCI bus was sufficient to
keep PV drivers binding. However, this does not protect against the
active device going away so you can still end up with a non-bootable VM.

This patch re-instates the code, modified to check for presence of the
active device with a wildcard instance rather than a fixed instance.

NOTE: This patch does include a change to emulated_interface.h, to allow
      a NULL instance ID to be passed to IsDevicePresent. Whilst this is
      a new semantic, it does not affect compatibility so there is no
      change to the interface version.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>