Add strict version check between XEN and XENFILT/XENBUS
Because the ABI between XEN and the other drivers in the package is
(intentionally) unstable, add a strict version check using the single
function XenTouch to prevent XENBUS or XENFILT loading if an incumbent
XEN is from a previous package installation.

Also add code to the co-installer to request a reboot, as this is needed
to bring up a compatible set of modules.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>