Revert commit 632cc904 "Remove PDO set/is-missing logic from...
... XENFILT" and re-work PnP code again.

In WHQL testing I suspect the removal and re-creation of filter objects
when IRP_MN_REMOVE_DEVICE is processed in the case that underlying PDO is
not actually going away may cause problems.

By reverting 632cc904 this bouncing is prevented but the code needs more
work to fix the hanging object references from filtDO to PDO that were the
motivation for 632cc904 in the first place.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>