monitor: get dialog paramaters from the registry
It is easier to localise the monitor dialog if it picks up the reboot dialog
title and message from registry parameters rather than having the hardcoded
or in a string table. This patch does this and sets default values in the
the INF file.

This patchs also adds a call to wait for driver installations to complete
before initialiating a reboot.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>