Add basic support for PV console
Xen toolstacks have, for some time, created a PV console even for HVM
guests but, so far, Windows has had no frontend for this.

This patch adds the basic plumbing for the PV console, under a new
interface called XENBUS_CONSOLE. This interface is currently private to
XENBUS (so not accessible by child drivers) and will initially only support
writing.
The patch adds a new log disposition that will write INFO, WARNING and
ERROR level messages, so these can now be seen by attaching to the console
backend in the toolstack domain.

Future patches will add read support and a new character device to make
the console available to user-space.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>