Get rid of old XenServer-ism
Old versions of XenServer used to require that Windows tell the hypervisor
whether it was 64-bit or 32-bit so that the shared info page could be layed
out correctly.

This requirement was dropped in later versions of XenServer but some
versions had a bug where the domain wallclock time was not updated
correctly unless the old mechanism was used.

All such versions of XenServer have long been out of support and such hacks
really have no place in the Xen Project code-base anyway.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>