Partially revert "Simplify EVTCHN Unmask"
This partially reverts commit 3ad02ccc6c68, which removed the possibility
that unmasking an event channel could fail. This semantic was actually
useful in some cases though, specifically to avoid dropping out of a DPC
only to have it immediately re-queued.

This patch, therefore, re-instates the older semantic but makes it optional
so that callers that cannot easily tolerate such a failure can force the
unmask to succeed.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>