Use new query method in XENBUS_GNTTAB method to get STORE and CONSOLE PFNs
Since XENBUS_GNTTAB now has a query method, the PFNs of the STORE and
CONSOLE shared pages can be found without the need to make hypercalls (to
get them via HVMOP_get_param).

This patch makes use of the new mechanism but this also requires a
re-ordering of some of the interface initialization code in FdoCreate().

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>