Fix event channel ABI fall-back
If the version of Xen that the guest is running on does not support the
FIFO event channel ABI then the code is supposed to fall back to using the
old '2-level' ABI through the shared info page. However, this fall-back
does not work correctly because the error path in EvtchnFifoAcquire()
causes the error status to be overwritten with STATUS_SUCCESS.

This patch makes sure the error status is preserved.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>