Sample 'RealTimeIsUniversal' at boot time...
...and report via the XENBUS_SHARED_INFO GetTime method

The DWORD registry value:

HKLM\System\CurrentControlSet\Control\TimeZoneInformation\RealTimeIsUniversal

determines whether Windows programs the RTC in local time or UTC. If the
value is zero or missing (which is the default) then the RTC will be
programmed in local time, otherwise it will be programmed in UTC.

Since the emulated RTC and the Xen wall-clock are kept in sync, it is
necessary to know which mode Windows is operating in to correctly interpret
the wall-clock value. This patch checks the registry value on boot (it is
necessary to reboot after modifying it for it to take effect) and reports
the RTC operating mode along-side the Xen wall-clock in a updated
GetTime method in version 3 of the XENBUS_SHARED_INFO interface.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>