Update XENFILT_EMULATED IsDiskPresent() method...
...to match how XENVBD actually uses it.

XENVBD erroneously passes the disk number of the PV disk as the target
number, which is incorrect for disk numbers more than 1. E.g. disk 3 should
equate to controller 1, target 1.
This patch adapts the method to that use by dropping the extraneous
Controller and Lun parameters and properly interpreting the Target
parameter.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>